# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EdgeGuard v3.0 API - Institutional Quant & Protection Engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Fixture {
  fixtureId: number;
  competition: string;
  homeTeam: string;
  awayTeam: string;
  kickoffTs: number;
  status: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  minutePlayed?: number | null;
  /** @nullable */
  currentEdgeScore?: number | null;
  /** @nullable */
  feedLatencyMs?: number | null;
}

export interface OddsSnapshot {
  id: number;
  fixtureId: number;
  ts: number;
  market: string;
  selection: string;
  stablePrice: number;
  spread: number;
  volume: number;
}

export type AgentSignalPayload = { [key: string]: unknown };

export interface AgentSignal {
  id: number;
  fixtureId: number;
  ts: number;
  agentName: string;
  signalType: string;
  confidence: number;
  payload?: AgentSignalPayload;
}

export interface AgentHeartbeat {
  agentName: string;
  status: string;
  /** @nullable */
  lastSignalTs?: number | null;
  activeFixtures: number;
  /** @nullable */
  verdict?: string | null;
}

export interface Alert {
  id: number;
  /** @nullable */
  userId?: string | null;
  fixtureId: number;
  ts: number;
  edgeScore: number;
  /** @nullable */
  narration?: string | null;
  action: string;
  fired: boolean;
  fixture?: Fixture;
}

export type ReceiptMemoJson = { [key: string]: unknown };

export interface Receipt {
  id: number;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  alertId?: number | null;
  /** @nullable */
  txSignature?: string | null;
  memoJson?: ReceiptMemoJson;
  cluster: string;
  ts: number;
  status: string;
  fixture?: Fixture;
  alert?: Alert;
}

export type ReceiptInputMemoJson = { [key: string]: unknown };

export interface ReceiptInput {
  /** @nullable */
  alertId?: number | null;
  fixtureId: number;
  txSignature: string;
  memoJson: ReceiptInputMemoJson;
  cluster: string;
  status: string;
}

export interface DashboardSummary {
  activeFixtures: number;
  totalAlerts: number;
  executedToday: number;
  vetoedToday: number;
  avgEdgeScore: number;
  feedLatencyMs: number;
  activeAgents: number;
  alertsPerHour: number;
}

export interface LiveTickerItem {
  fixtureId: number;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  minutePlayed: number;
  edgeScore: number;
  action: string;
  latencyMs: number;
  /** @nullable */
  topSignal?: string | null;
}

export interface RiskGridItem {
  fixtureId: number;
  homeTeam: string;
  awayTeam: string;
  spread: number;
  latencyMs: number;
  volatilityRisk: string;
  sentinelStatus: string;
  edgeScore: number;
  recommendation: string;
}

export type ScoreEventMeta = { [key: string]: unknown };

export interface ScoreEvent {
  id: number;
  fixtureId: number;
  ts: number;
  minute: number;
  eventType: string;
  participant: string;
  meta?: ScoreEventMeta;
}

export interface FixtureTimeline {
  fixtureId: number;
  odds: OddsSnapshot[];
  signals: AgentSignal[];
  alerts: Alert[];
  scoreEvents: ScoreEvent[];
}

export interface TxlineActivateInput {
  walletAddress: string;
  signature: string;
  message: string;
  network: string;
}

export interface TxlineSession {
  apiToken: string;
  network: string;
  expiresAt: number;
}

export type NarrateInputAgentSignalsItem = { [key: string]: unknown };

export type NarrateInputContext = { [key: string]: unknown };

export interface NarrateInput {
  alertId: number;
  fixtureId: number;
  edgeScore: number;
  action: string;
  agentSignals?: NarrateInputAgentSignalsItem[];
  context?: NarrateInputContext;
}

export interface NarrateResult {
  narration: string;
  source: string;
}

export type ListFixturesParams = {
status?: string;
limit?: number;
};

export type ListAgentSignalsParams = {
agentName?: string;
signalType?: string;
limit?: number;
};

export type ListAlertsParams = {
fixtureId?: number;
action?: string;
limit?: number;
};

export type ListReceiptsParams = {
status?: string;
limit?: number;
};