# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EdgeGuard v3.0 API - Institutional Quant & Protection Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fixtures:
    get:
      operationId: listFixtures
      tags: [fixtures]
      summary: List all tracked fixtures
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of fixtures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fixture"
  /fixtures/{fixtureId}:
    get:
      operationId: getFixture
      tags: [fixtures]
      summary: Get fixture by ID
      parameters:
        - name: fixtureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fixture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fixtures/{fixtureId}/odds:
    get:
      operationId: getFixtureOdds
      tags: [odds]
      summary: Get odds snapshots for a fixture
      parameters:
        - name: fixtureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Odds snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OddsSnapshot"
  /fixtures/{fixtureId}/agents:
    get:
      operationId: getFixtureAgentSignals
      tags: [agents]
      summary: Get agent signals for a fixture
      parameters:
        - name: fixtureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSignal"
  /fixtures/{fixtureId}/timeline:
    get:
      operationId: getFixtureTimeline
      tags: [fixtures]
      summary: Get full timeline replay data for a fixture
      parameters:
        - name: fixtureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixtureTimeline"
  /agents/signals:
    get:
      operationId: listAgentSignals
      tags: [agents]
      summary: List recent agent signals across all fixtures
      parameters:
        - name: agentName
          in: query
          schema:
            type: string
        - name: signalType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Agent signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSignal"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts for the authenticated user
      parameters:
        - name: fixtureId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{alertId}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert by ID
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
  /narrate:
    post:
      operationId: narrateAlert
      tags: [narrate]
      summary: Generate AI narration for an alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NarrateInput"
      responses:
        "200":
          description: Narration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NarrateResult"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Fixture:
      type: object
      properties:
        fixtureId:
          type: integer
        competition:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        kickoffTs:
          type: integer
        status:
          type: string
        homeScore:
          type: ["integer", "null"]
        awayScore:
          type: ["integer", "null"]
        minutePlayed:
          type: ["integer", "null"]
        currentEdgeScore:
          type: ["integer", "null"]
        feedLatencyMs:
          type: ["integer", "null"]
      required:
        - fixtureId
        - competition
        - homeTeam
        - awayTeam
        - kickoffTs
        - status
    OddsSnapshot:
      type: object
      properties:
        id:
          type: integer
        fixtureId:
          type: integer
        ts:
          type: integer
        market:
          type: string
        selection:
          type: string
        stablePrice:
          type: number
        spread:
          type: number
        volume:
          type: number
      required:
        - id
        - fixtureId
        - ts
        - market
        - selection
        - stablePrice
        - spread
        - volume
    AgentSignal:
      type: object
      properties:
        id:
          type: integer
        fixtureId:
          type: integer
        ts:
          type: integer
        agentName:
          type: string
        signalType:
          type: string
        confidence:
          type: number
        payload:
          type: object
          additionalProperties: true
      required:
        - id
        - fixtureId
        - ts
        - agentName
        - signalType
        - confidence
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        fixtureId:
          type: integer
        ts:
          type: integer
        edgeScore:
          type: integer
        narration:
          type: ["string", "null"]
        action:
          type: string
        fired:
          type: boolean
        fixture:
          $ref: "#/components/schemas/Fixture"
      required:
        - id
        - fixtureId
        - ts
        - edgeScore
        - action
        - fired
...
    FixtureTimeline:
      type: object
      properties:
        fixtureId:
          type: integer
        odds:
          type: array
          items:
            $ref: "#/components/schemas/OddsSnapshot"
        signals:
          type: array
          items:
            $ref: "#/components/schemas/AgentSignal"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        scoreEvents:
          type: array
          items:
            $ref: "#/components/schemas/ScoreEvent"
      required:
        - fixtureId
        - odds
        - signals
        - alerts
        - scoreEvents
    ScoreEvent:
      type: object
      properties:
        id:
          type: integer
        fixtureId:
          type: integer
        ts:
          type: integer
        minute:
          type: integer
        eventType:
          type: string
        participant:
          type: string
        meta:
          type: object
          additionalProperties: true
      required:
        - id
        - fixtureId
        - ts
        - minute
        - eventType
        - participant
...
    NarrateInput:
      type: object
      properties:
        alertId:
          type: integer
        fixtureId:
          type: integer
        edgeScore:
          type: integer
        action:
          type: string
        agentSignals:
          type: array
          items:
            type: object
            additionalProperties: true
        context:
          type: object
          additionalProperties: true
      required:
        - alertId
        - fixtureId
        - edgeScore
        - action
    NarrateResult:
      type: object
      properties:
        narration:
          type: string
        source:
          type: string
      required:
        - narration
        - source