# lib/api-zod/src/generated/types/fixtureTimeline.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EdgeGuard v3.0 API - Institutional Quant & Protection Engine
 * OpenAPI spec version: 0.1.0
 */
import type { AgentSignal } from './agentSignal';
import type { Alert } from './alert';
import type { OddsSnapshot } from './oddsSnapshot';
import type { ScoreEvent } from './scoreEvent';

export interface FixtureTimeline {
  fixtureId: number;
  odds: OddsSnapshot[];
  signals: AgentSignal[];
  alerts: Alert[];
  scoreEvents: ScoreEvent[];
}