# lib/api-zod/src/generated/types/narrateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EdgeGuard v3.0 API - Institutional Quant & Protection Engine
 * OpenAPI spec version: 0.1.0
 */
import type { NarrateInputAgentSignalsItem } from './narrateInputAgentSignalsItem';
import type { NarrateInputContext } from './narrateInputContext';

export interface NarrateInput {
  alertId: number;
  fixtureId: number;
  edgeScore: number;
  action: string;
  agentSignals?: NarrateInputAgentSignalsItem[];
  context?: NarrateInputContext;
}