# lib/api-zod/src/generated/types/narrateResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EdgeGuard v3.0 API - Institutional Quant & Protection Engine
 * OpenAPI spec version: 0.1.0
 */

export interface NarrateResult {
  narration: string;
  source: string;
}