# lib/api-zod/src/generated/types/oddsSnapshot.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EdgeGuard v3.0 API - Institutional Quant & Protection Engine
 * OpenAPI spec version: 0.1.0
 */

export interface OddsSnapshot {
  id: number;
  fixtureId: number;
  ts: number;
  market: string;
  selection: string;
  stablePrice: number;
  spread: number;
  volume: number;
}